requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==24.4.2
bleach==4.1.0
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
chitra==0.2.0
click==8.1.3
contourpy==1.0.6
cryptography==38.0.1
cv==1.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docker==6.1.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.12.0
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2023.6.0
gmpy2==2.1.2
idna==3.4
imageio==2.31.1
imgaug==0.4.0
ipykernel==6.15.2
ipython==8.6.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.6
jsonschema==4.16.0
jupyter_client==7.4.7
jupyter_core==4.11.2
jupyter-server==1.18.1
jupyterlab==3.5.0
jupyterlab-pygments==0.1.2
jupyterlab_server==2.16.3
kiwisolver==1.4.4
lazy_loader==0.3
lightning-utilities==0.9.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==3.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.3
online-triplet-loss==0.0.6
opencv-python==4.6.0.66
packaging==23.0
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.2.2
platformdirs==3.1.0
pooch==1.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.20
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pycocotools==2.0.7
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
scikit-image==0.21.0
scikit-learn>=1.3.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.5.0
shapely==2.0.1
six==1.16.0
sniffio==1.2.0
soupsieve>=2.3.2
stack-data==0.2.0
sympy==1.12
terminado==0.13.1
threadpoolctl==2.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
torch~=2.1.0
torchaudio~=2.1.0
torchmetrics==1.1.0
torchvision~=0.16.0
tornado==6.2
tqdm==4.65.0
traitlets==5.1.1
typeguard==4.1.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
yarl==1.9.4
lightning